fix: harden PATCH recovery and tombstones
