+++ crates/temper-server/tests/dst_persistence.rs
@@ -7,21 +7,23 @@
 //!
 //! FoundationDB pattern: same code, simulated I/O.
 
-use std::collections::{BTreeMap, VecDeque};
-use std::sync::{Arc, Mutex, RwLock};
+#[path = "dst_persistence/field_update_durability.rs"]
+mod field_update_durability;
+#[path = "dst_persistence/field_update_recovery.rs"]
+mod field_update_recovery;
+#[path = "dst_persistence/field_update_support.rs"]
+mod field_update_support;
+
+use std::collections::BTreeMap;
+use std::sync::{Arc, RwLock};
 use std::time::Duration;
 
 use temper_jit::table::TransitionTable;
 use temper_runtime::ActorSystem;
-use temper_runtime::persistence::{
-    EntityKeyRow, EntityVectorRow, EventMetadata, EventStore, IndexReconciliation,
-    PersistenceAppend, PersistenceAppendResult, PersistenceEnvelope, PersistenceError,
-};
-use temper_runtime::scheduler::{install_deterministic_context, sim_now, sim_uuid};
-use temper_server::key_index::canonical_key_hash;
+use temper_runtime::scheduler::install_deterministic_context;
 use temper_server::storage::{BackendLabel, BoxedEventStore};
 use temper_server::{EntityActor, EntityMsg, EntityResponse};
-use temper_store_sim::{SimEventStore, SimFaultConfig};
+use temper_store_sim::SimEventStore;
 
 const ORDER_IOA: &str = include_str!("../../../test-fixtures/specs/order.ioa.toml");
 const NUM_SEEDS: u64 = 100;
@@ -32,170 +34,6 @@ fn order_table() -> Arc<RwLock<TransitionTable>> {
 
 fn sim_store(seed: u64) -> BoxedEventStore {
     BoxedEventStore::new(SimEventStore::no_faults(seed))
-}
-
-#[derive(Clone)]
-struct ConflictBeforeAppendStore {
-    inner: SimEventStore,
-    conflicts: Arc<Mutex<BTreeMap<String, VecDeque<PersistenceEnvelope>>>>,
-}
-
-impl ConflictBeforeAppendStore {
-    fn new(seed: u64) -> Self {
-        Self {
-            inner: SimEventStore::no_faults(seed),
-            conflicts: Arc::new(Mutex::new(BTreeMap::new())),
-        }
-    }
-
-    fn queue_conflicts(&self, persistence_id: &str, events: Vec<PersistenceEnvelope>) {
-        self.conflicts
-            .lock()
-            .expect("conflict script lock poisoned")
-            .insert(persistence_id.to_string(), events.into());
-    }
-
-    fn scripted_conflict(&self, persistence_id: &str) -> Option<PersistenceEnvelope> {
-        let mut conflicts = self
-            .conflicts
-            .lock()
-            .expect("conflict script lock poisoned");
-        let event = conflicts.get_mut(persistence_id)?.pop_front();
-        if conflicts
-            .get(persistence_id)
-            .is_some_and(VecDeque::is_empty)
-        {
-            conflicts.remove(persistence_id);
-        }
-        event
-    }
-}
-
-impl EventStore for ConflictBeforeAppendStore {
-    async fn append(
-        &self,
-        persistence_id: &str,
-        expected_sequence: u64,
-        events: &[PersistenceEnvelope],
-    ) -> Result<u64, PersistenceError> {
-        self.inner
-            .append(persistence_id, expected_sequence, events)
-            .await
-    }
-
-    async fn append_with_index_rows(
-        &self,
-        persistence_id: &str,
-        expected_sequence: u64,
-        events: &[PersistenceEnvelope],
-        key_rows: &[EntityKeyRow],
-        vector_rows: &[EntityVectorRow],
-        reconciliation: IndexReconciliation,
-    ) -> Result<u64, PersistenceError> {
-        if let Some(conflict) = self.scripted_conflict(persistence_id) {
-            let actual = self
-                .inner
-                .append(persistence_id, expected_sequence, &[conflict])
-                .await?;
-            return Err(PersistenceError::ConcurrencyViolation {
-                expected: expected_sequence,
-                actual,
-            });
-        }
-        self.inner
-            .append_with_index_rows(
-                persistence_id,
-                expected_sequence,
-                events,
-                key_rows,
-                vector_rows,
-                reconciliation,
-            )
-            .await
-    }
-
-    async fn append_batch(
-        &self,
-        appends: &[PersistenceAppend],
-    ) -> Result<Vec<PersistenceAppendResult>, PersistenceError> {
-        self.inner.append_batch(appends).await
-    }
-
-    async fn read_events(
-        &self,
-        persistence_id: &str,
-        from_sequence: u64,
-    ) -> Result<Vec<PersistenceEnvelope>, PersistenceError> {
-        self.inner.read_events(persistence_id, from_sequence).await
-    }
-
-    async fn save_snapshot(
-        &self,
-        persistence_id: &str,
-        sequence_nr: u64,
-        snapshot: &[u8],
-    ) -> Result<(), PersistenceError> {
-        self.inner
-            .save_snapshot(persistence_id, sequence_nr, snapshot)
-            .await
-    }
-
-    async fn load_snapshot(
-        &self,
-        persistence_id: &str,
-    ) -> Result<Option<(u64, Vec<u8>)>, PersistenceError> {
-        self.inner.load_snapshot(persistence_id).await
-    }
-
-    async fn list_entity_ids(
-        &self,
-        tenant: &str,
-    ) -> Result<Vec<(String, String)>, PersistenceError> {
-        self.inner.list_entity_ids(tenant).await
-    }
-
-    async fn list_entity_ids_by_type(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-    ) -> Result<Vec<String>, PersistenceError> {
-        self.inner
-            .list_entity_ids_by_type(tenant, entity_type)
-            .await
-    }
-}
-
-fn concurrent_add_item(persistence_id: &str, product_id: &str) -> PersistenceEnvelope {
-    PersistenceEnvelope {
-        sequence_nr: 0,
-        event_type: "AddItem".to_string(),
-        payload: serde_json::json!({
-            "action": "AddItem",
-            "from_status": "Draft",
-            "to_status": "Draft",
-            "timestamp": sim_now(),
-            "params": {"ProductId": product_id}
-        }),
-        metadata: EventMetadata {
-            event_id: sim_uuid(),
-            causation_id: sim_uuid(),
-            correlation_id: sim_uuid(),
-            timestamp: sim_now(),
-            actor_id: persistence_id.to_string(),
-        },
-    }
-}
-
-fn order_key_hash(workspace: &str, path: &str) -> String {
-    canonical_key_hash(
-        "ws_path",
-        &["WorkspaceId".to_string(), "Path".to_string()],
-        &serde_json::Map::from_iter([
-            ("WorkspaceId".to_string(), serde_json::json!(workspace)),
-            ("Path".to_string(), serde_json::json!(path)),
-        ]),
-    )
-    .expect("complete order key")
 }
 
 async fn dispatch_action(
@@ -220,20 +58,6 @@ async fn dispatch_action(
 async fn get_state(actor_ref: &temper_runtime::actor::ActorRef<EntityMsg>) -> EntityResponse {
     actor_ref
         .ask(EntityMsg::GetState, Duration::from_secs(5))
-        .await
-        .expect("actor should respond")
-}
-
-async fn update_fields(
-    actor_ref: &temper_runtime::actor::ActorRef<EntityMsg>,
-    fields: serde_json::Value,
-    replace: bool,
-) -> EntityResponse {
-    actor_ref
-        .ask(
-            EntityMsg::UpdateFields { fields, replace },
-            Duration::from_secs(5),
-        )
         .await
         .expect("actor should respond")
 }
@@ -581,676 +405,4 @@ async fn dst_replay_preserves_data_fields() {
             "seed {seed}: fields mismatch after replay"
         );
     }
-}
-
-// =========================================================================
-// Regression: PATCH-only fields survive actor replacement
-// =========================================================================
-
-#[tokio::test]
-async fn dst_patch_only_fields_survive_actor_replacement() {
-    for seed in 0..NUM_SEEDS {
-        let (_guard, _clock, _id_gen) = install_deterministic_context(seed);
-        let store = sim_store(seed);
-        let table = order_table();
-        let entity_id = format!("ord-patch-only-{seed}");
-
-        {
-            let system = ActorSystem::new("dst-patch-only-1");
-            let actor = EntityActor::with_persistence(
-                "Order",
-                &entity_id,
-                table.clone(),
-                serde_json::json!({
-                    "Title": "before-patch",
-                    "StableField": "preserved",
-                    "WorkspaceId": "ws-patch",
-                    "Path": "/before"
-                }),
-                store.clone(),
-                BackendLabel::Sim,
-            )
-            .with_tenant("default");
-            let actor_ref = system.spawn(actor, &entity_id);
-
-            let patched = update_fields(
-                &actor_ref,
-                serde_json::json!({
-                    "Title": "after-patch",
-                    "Priority": "High",
-                    "Path": "/after"
-                }),
-                false,
-            )
-            .await;
-            assert!(
-                patched.success,
-                "seed {seed}: PATCH failed: {:?}",
-                patched.error
-            );
-            assert_eq!(patched.state.fields["Title"], "after-patch");
-            assert_eq!(patched.state.fields["Priority"], "High");
-            assert_eq!(
-                store
-                    .lookup_by_key(
-                        "default",
-                        "Order",
-                        "ws_path",
-                        &order_key_hash("ws-patch", "/before"),
-                    )
-                    .await
-                    .expect("old key lookup should succeed"),
-                None,
-                "seed {seed}: old declared-key projection was not removed"
-            );
-            assert_eq!(
-                store
-                    .lookup_by_key(
-                        "default",
-                        "Order",
-                        "ws_path",
-                        &order_key_hash("ws-patch", "/after"),
-                    )
-                    .await
-                    .expect("new key lookup should succeed"),
-                Some(entity_id.clone()),
-                "seed {seed}: PATCH did not co-commit its declared-key projection"
-            );
-        }
-
-        {
-            let (_guard2, _clock2, _id_gen2) = install_deterministic_context(seed + 10_000);
-            let system = ActorSystem::new("dst-patch-only-2");
-            let actor = EntityActor::with_persistence(
-                "Order",
-                &entity_id,
-                table.clone(),
-                serde_json::json!({}),
-                store.clone(),
-                BackendLabel::Sim,
-            )
-            .with_tenant("default");
-            let actor_ref = system.spawn(actor, format!("{entity_id}-replacement"));
-
-            let replayed = get_state(&actor_ref).await;
-            assert_eq!(
-                replayed.state.fields["Title"], "after-patch",
-                "seed {seed}: acknowledged PATCH was lost after actor replacement"
-            );
-            assert_eq!(
-                replayed.state.fields["Priority"], "High",
-                "seed {seed}: PATCH-only field was lost after actor replacement"
-            );
-            assert_eq!(replayed.state.fields["StableField"], "preserved");
-
-            let replaced =
-                update_fields(&actor_ref, serde_json::json!({"Title": "after-put"}), true).await;
-            assert!(replaced.success, "seed {seed}: PUT failed");
-            assert!(replaced.state.fields.get("Priority").is_none());
-            assert!(replaced.state.fields.get("StableField").is_none());
-            assert_eq!(
-                store
-                    .lookup_by_key(
-                        "default",
-                        "Order",
-                        "ws_path",
-                        &order_key_hash("ws-patch", "/after"),
-                    )
-                    .await
-                    .expect("removed key lookup should succeed"),
-                None,
-                "seed {seed}: PUT left a stale declared-key projection"
-            );
-        }
-
-        let (_guard3, _clock3, _id_gen3) = install_deterministic_context(seed + 20_000);
-        let system = ActorSystem::new("dst-patch-only-3");
-        let actor = EntityActor::with_persistence(
-            "Order",
-            &entity_id,
-            table.clone(),
-            serde_json::json!({}),
-            store.clone(),
-            BackendLabel::Sim,
-        )
-        .with_tenant("default");
-        let actor_ref = system.spawn(actor, format!("{entity_id}-put-replacement"));
-        let replayed = get_state(&actor_ref).await;
-        assert_eq!(replayed.state.fields["Title"], "after-put");
-        assert!(
-            replayed.state.fields.get("Priority").is_none(),
-            "seed {seed}: PUT replay resurrected a removed PATCH field"
-        );
-        assert!(replayed.state.fields.get("StableField").is_none());
-        assert_eq!(
-            store
-                .lookup_by_key(
-                    "default",
-                    "Order",
-                    "ws_path",
-                    &order_key_hash("ws-patch", "/after"),
-                )
-                .await
-                .expect("replayed removed key lookup should succeed"),
-            None,
-            "seed {seed}: stale key returned after PUT replay"
-        );
-    }
-}
-
-// =========================================================================
-// Regression: a failed journal append cannot publish a field update
-// =========================================================================
-
-#[tokio::test]
-async fn dst_field_update_fails_closed_when_journal_append_fails() {
-    let (_guard, _clock, _id_gen) = install_deterministic_context(189);
-    let store_inner = SimEventStore::no_faults(189);
-    let store = BoxedEventStore::new(store_inner.clone());
-    let table = order_table();
-    let system = ActorSystem::new("dst-field-update-fail-closed");
-    let actor = EntityActor::with_persistence(
-        "Order",
-        "ord-field-failure",
-        table,
-        serde_json::json!({"Title": "durable-before"}),
-        store,
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let actor_ref = system.spawn(actor, "ord-field-failure");
-
-    let before = get_state(&actor_ref).await;
-    assert_eq!(before.state.fields["Title"], "durable-before");
-    let sequence_before = before.state.sequence_nr;
-
-    store_inner.restore_faults(SimFaultConfig {
-        write_failure_prob: 1.0,
-        ..SimFaultConfig::none()
-    });
-
-    let response = update_fields(
-        &actor_ref,
-        serde_json::json!({"Title": "volatile-after"}),
-        false,
-    )
-    .await;
-
-    assert!(
-        !response.success,
-        "field update must not report success when its journal append fails"
-    );
-    assert_eq!(response.state.fields["Title"], "durable-before");
-    assert_eq!(response.state.sequence_nr, sequence_before);
-
-    let live = get_state(&actor_ref).await;
-    assert_eq!(live.state.fields["Title"], "durable-before");
-    assert_eq!(live.state.sequence_nr, sequence_before);
-}
-
-#[tokio::test]
-async fn dst_field_update_retries_after_concurrency_violation() {
-    let (_guard, _clock, _id_gen) = install_deterministic_context(18_900);
-    let store_inner = SimEventStore::no_faults(18_900);
-    let store = BoxedEventStore::new(store_inner.clone());
-    let table = order_table();
-    let entity_id = "ord-field-concurrency";
-    let persistence_id = format!("default:Order:{entity_id}");
-
-    {
-        let system = ActorSystem::new("dst-field-update-concurrency-1");
-        let actor = EntityActor::with_persistence(
-            "Order",
-            entity_id,
-            table.clone(),
-            serde_json::json!({"Title": "before-retry"}),
-            store.clone(),
-            BackendLabel::Sim,
-        )
-        .with_tenant("default");
-        let actor_ref = system.spawn(actor, entity_id);
-        let before = get_state(&actor_ref).await;
-        assert_eq!(before.state.sequence_nr, 1);
-
-        store_inner
-            .append(
-                &persistence_id,
-                1,
-                &[PersistenceEnvelope {
-                    sequence_nr: 2,
-                    event_type: "AddItem".to_string(),
-                    payload: serde_json::json!({
-                        "action": "AddItem",
-                        "from_status": "Draft",
-                        "to_status": "Draft",
-                        "timestamp": sim_now(),
-                        "params": {"ProductId": "concurrent-item"}
-                    }),
-                    metadata: EventMetadata {
-                        event_id: sim_uuid(),
-                        causation_id: sim_uuid(),
-                        correlation_id: sim_uuid(),
-                        timestamp: sim_now(),
-                        actor_id: persistence_id.clone(),
-                    },
-                }],
-            )
-            .await
-            .expect("concurrent action should advance authoritative history");
-        let response = update_fields(
-            &actor_ref,
-            serde_json::json!({"Title": "after-retry"}),
-            false,
-        )
-        .await;
-        assert!(
-            response.success,
-            "field update retry failed: {:?}",
-            response.error
-        );
-        assert_eq!(response.state.item_count, 1);
-        assert_eq!(response.state.fields["ProductId"], "concurrent-item");
-    }
-
-    let system = ActorSystem::new("dst-field-update-concurrency-2");
-    let actor = EntityActor::with_persistence(
-        "Order",
-        entity_id,
-        table,
-        serde_json::json!({}),
-        store,
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let actor_ref = system.spawn(actor, "ord-field-concurrency-replacement");
-    let replayed = get_state(&actor_ref).await;
-    assert_eq!(replayed.state.fields["Title"], "after-retry");
-    assert_eq!(replayed.state.item_count, 1);
-    assert_eq!(replayed.state.fields["ProductId"], "concurrent-item");
-}
-
-#[tokio::test]
-async fn dst_field_update_retry_exhaustion_is_reported_distinctly() {
-    let (_guard, _clock, _id_gen) = install_deterministic_context(18_902);
-    let store_inner = SimEventStore::no_faults(18_902);
-    let store = BoxedEventStore::new(store_inner.clone());
-    let table = order_table();
-    let entity_id = "ord-field-concurrency-exhausted";
-    let persistence_id = format!("default:Order:{entity_id}");
-    let system = ActorSystem::new("dst-field-update-concurrency-exhausted");
-    let actor = EntityActor::with_persistence(
-        "Order",
-        entity_id,
-        table,
-        serde_json::json!({"Title": "durable-before"}),
-        store,
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let actor_ref = system.spawn(actor, entity_id);
-
-    let before = get_state(&actor_ref).await;
-    assert_eq!(before.state.sequence_nr, 1);
-    store_inner.inject_concurrency_violations(&persistence_id, 4);
-
-    let response = update_fields(
-        &actor_ref,
-        serde_json::json!({"Title": "must-not-publish"}),
-        false,
-    )
-    .await;
-
-    assert!(!response.success, "exhausted retries must fail the caller");
-    assert_eq!(
-        response.error.as_deref(),
-        Some("field update retry budget exhausted"),
-        "retry exhaustion must remain distinguishable from a non-concurrency persistence failure"
-    );
-    assert_eq!(response.state.fields["Title"], "durable-before");
-    assert_eq!(response.state.sequence_nr, before.state.sequence_nr);
-    assert_eq!(
-        store_inner.pending_concurrency_violations(&persistence_id),
-        1,
-        "one initial attempt plus two retries must consume exactly three violations"
-    );
-    assert_eq!(
-        store_inner.dump_journal(&persistence_id).len(),
-        1,
-        "failed field updates must not append or publish speculative state"
-    );
-}
-
-#[tokio::test]
-async fn dst_field_update_retry_exhaustion_keeps_recovered_history() {
-    let (_guard, _clock, _id_gen) = install_deterministic_context(18_903);
-    let store_inner = SimEventStore::no_faults(18_903);
-    let store = BoxedEventStore::new(store_inner.clone());
-    let table = order_table();
-    let entity_id = "ord-field-recovered-on-exhaustion";
-    let persistence_id = format!("default:Order:{entity_id}");
-    let system = ActorSystem::new("dst-field-recovered-on-exhaustion");
-    let actor = EntityActor::with_persistence(
-        "Order",
-        entity_id,
-        table,
-        serde_json::json!({"Title": "durable-before"}),
-        store,
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let actor_ref = system.spawn(actor, entity_id);
-
-    let before = get_state(&actor_ref).await;
-    assert_eq!(before.state.sequence_nr, 1);
-    store_inner
-        .append(
-            &persistence_id,
-            1,
-            &[PersistenceEnvelope {
-                sequence_nr: 2,
-                event_type: "AddItem".to_string(),
-                payload: serde_json::json!({
-                    "action": "AddItem",
-                    "from_status": "Draft",
-                    "to_status": "Draft",
-                    "timestamp": sim_now(),
-                    "params": {"ProductId": "concurrent-item"}
-                }),
-                metadata: EventMetadata {
-                    event_id: sim_uuid(),
-                    causation_id: sim_uuid(),
-                    correlation_id: sim_uuid(),
-                    timestamp: sim_now(),
-                    actor_id: persistence_id.clone(),
-                },
-            }],
-        )
-        .await
-        .expect("concurrent action should advance authoritative history");
-    store_inner.inject_concurrency_violations(&persistence_id, 3);
-
-    let response = update_fields(
-        &actor_ref,
-        serde_json::json!({"Title": "must-not-publish"}),
-        false,
-    )
-    .await;
-
-    assert!(!response.success, "exhausted retries must fail the PATCH");
-    assert_eq!(
-        response.error.as_deref(),
-        Some("field update retry budget exhausted")
-    );
-    assert_eq!(
-        response.state.sequence_nr, 2,
-        "the actor must retain the authoritative sequence recovered during retry"
-    );
-    assert_eq!(response.state.item_count, 1);
-    assert_eq!(response.state.fields["ProductId"], "concurrent-item");
-    assert_eq!(response.state.fields["Title"], "durable-before");
-
-    let live = get_state(&actor_ref).await;
-    assert_eq!(live.state.sequence_nr, 2);
-    assert_eq!(live.state.item_count, 1);
-    assert_eq!(live.state.fields["ProductId"], "concurrent-item");
-    assert_eq!(live.state.fields["Title"], "durable-before");
-    assert_eq!(
-        store_inner.dump_journal(&persistence_id).len(),
-        2,
-        "the failed PATCH must not append speculative history"
-    );
-}
-
-#[tokio::test]
-async fn dst_field_update_retry_exhaustion_recovers_the_final_real_writer() {
-    let seed = 18_904;
-    let (_guard, _clock, _id_gen) = install_deterministic_context(seed);
-    let scripted_store = ConflictBeforeAppendStore::new(seed);
-    let store_inner = scripted_store.inner.clone();
-    let store = BoxedEventStore::new(scripted_store.clone());
-    let table = order_table();
-    let entity_id = "ord-field-final-real-writer";
-    let persistence_id = format!("default:Order:{entity_id}");
-    let system = ActorSystem::new("dst-field-final-real-writer");
-    let actor = EntityActor::with_persistence(
-        "Order",
-        entity_id,
-        table,
-        serde_json::json!({"Title": "durable-before"}),
-        store,
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let actor_ref = system.spawn(actor, entity_id);
-
-    let before = get_state(&actor_ref).await;
-    assert_eq!(before.state.sequence_nr, 1);
-    scripted_store.queue_conflicts(
-        &persistence_id,
-        vec![
-            concurrent_add_item(&persistence_id, "concurrent-item-1"),
-            concurrent_add_item(&persistence_id, "concurrent-item-2"),
-            concurrent_add_item(&persistence_id, "concurrent-item-3"),
-        ],
-    );
-
-    let response = update_fields(
-        &actor_ref,
-        serde_json::json!({"Title": "must-not-publish"}),
-        false,
-    )
-    .await;
-
-    assert!(!response.success, "exhausted retries must fail the PATCH");
-    assert_eq!(
-        response.error.as_deref(),
-        Some("field update retry budget exhausted")
-    );
-    assert_eq!(
-        response.state.sequence_nr, 4,
-        "the exhausted response must include the writer that won the final attempt"
-    );
-    assert_eq!(response.state.item_count, 3);
-    assert_eq!(response.state.fields["ProductId"], "concurrent-item-3");
-    assert_eq!(response.state.fields["Title"], "durable-before");
-
-    let live = get_state(&actor_ref).await;
-    assert_eq!(live.state.sequence_nr, 4);
-    assert_eq!(live.state.item_count, 3);
-    assert_eq!(live.state.fields["ProductId"], "concurrent-item-3");
-    assert_eq!(live.state.fields["Title"], "durable-before");
-    assert_eq!(
-        store_inner.dump_journal(&persistence_id).len(),
-        4,
-        "only the three real concurrent writers may follow Created"
-    );
-}
-
-#[tokio::test]
-async fn dst_field_update_recovery_read_failure_restarts_before_serving_state() {
-    let seed = 18_905;
-    let (_guard, _clock, _id_gen) = install_deterministic_context(seed);
-    let scripted_store = ConflictBeforeAppendStore::new(seed);
-    let store_inner = scripted_store.inner.clone();
-    let store = BoxedEventStore::new(scripted_store.clone());
-    let table = order_table();
-    let entity_id = "ord-field-recovery-read-failure";
-    let persistence_id = format!("default:Order:{entity_id}");
-    let system = ActorSystem::new("dst-field-recovery-read-failure");
-    let actor = EntityActor::with_persistence(
-        "Order",
-        entity_id,
-        table,
-        serde_json::json!({"Title": "durable-before"}),
-        store,
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let actor_ref = system.spawn(actor, entity_id);
-
-    let before = get_state(&actor_ref).await;
-    assert_eq!(before.state.sequence_nr, 1);
-    scripted_store.queue_conflicts(
-        &persistence_id,
-        vec![concurrent_add_item(
-            &persistence_id,
-            "concurrent-item-after-read-failure",
-        )],
-    );
-    store_inner.fail_next_reads(&persistence_id, 1);
-
-    let update = actor_ref
-        .ask::<EntityResponse>(
-            EntityMsg::UpdateFields {
-                fields: serde_json::json!({"Title": "must-not-publish"}),
-                replace: false,
-            },
-            Duration::from_secs(5),
-        )
-        .await;
-    assert!(
-        update.is_err(),
-        "failed authoritative recovery must fail the actor message so stale state cannot remain live"
-    );
-
-    let recovered = get_state(&actor_ref).await;
-    assert_eq!(
-        recovered.state.sequence_nr, 2,
-        "the restarted actor must replay the real writer before serving state"
-    );
-    assert_eq!(recovered.state.item_count, 1);
-    assert_eq!(
-        recovered.state.fields["ProductId"],
-        "concurrent-item-after-read-failure"
-    );
-    assert_eq!(recovered.state.fields["Title"], "durable-before");
-    assert_eq!(store_inner.dump_journal(&persistence_id).len(), 2);
-}
-
-#[tokio::test]
-async fn dst_field_update_restart_rejects_a_successful_looking_journal_prefix() {
-    let seed = 18_906;
-    let (_guard, _clock, _id_gen) = install_deterministic_context(seed);
-    let scripted_store = ConflictBeforeAppendStore::new(seed);
-    let store_inner = scripted_store.inner.clone();
-    let store = BoxedEventStore::new(scripted_store.clone());
-    let table = order_table();
-    let entity_id = "ord-field-restart-truncated-tail";
-    let persistence_id = format!("default:Order:{entity_id}");
-    let system = ActorSystem::new("dst-field-restart-truncated-tail");
-    let actor = EntityActor::with_persistence(
-        "Order",
-        entity_id,
-        table.clone(),
-        serde_json::json!({"Title": "durable-before"}),
-        store.clone(),
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let actor_ref = system.spawn(actor, entity_id);
-
-    let before = get_state(&actor_ref).await;
-    assert_eq!(before.state.sequence_nr, 1);
-    scripted_store.queue_conflicts(
-        &persistence_id,
-        vec![concurrent_add_item(
-            &persistence_id,
-            "concurrent-item-before-prefix",
-        )],
-    );
-    store_inner.restore_faults(SimFaultConfig {
-        read_truncation_prob: 1.0,
-        ..SimFaultConfig::none()
-    });
-    store_inner.fail_next_reads(&persistence_id, 1);
-
-    let update = actor_ref
-        .ask::<EntityResponse>(
-            EntityMsg::UpdateFields {
-                fields: serde_json::json!({"Title": "must-not-publish"}),
-                replace: false,
-            },
-            Duration::from_secs(5),
-        )
-        .await;
-    assert!(
-        update.is_err(),
-        "the explicit recovery failure must enter actor supervision"
-    );
-
-    let stale = actor_ref
-        .ask::<EntityResponse>(EntityMsg::GetState, Duration::from_secs(5))
-        .await;
-    assert!(
-        stale.is_err(),
-        "supervision must not publish a successful-looking prefix as current state"
-    );
-
-    store_inner.disable_faults();
-    let replacement = EntityActor::with_persistence(
-        "Order",
-        entity_id,
-        table,
-        serde_json::json!({"Title": "durable-before"}),
-        store,
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let replacement_ref = system.spawn(replacement, "replacement-after-truncated-tail");
-    let recovered = get_state(&replacement_ref).await;
-    assert_eq!(recovered.state.sequence_nr, 2);
-    assert_eq!(recovered.state.item_count, 1);
-    assert_eq!(
-        recovered.state.fields["ProductId"],
-        "concurrent-item-before-prefix"
-    );
-    assert_eq!(recovered.state.fields["Title"], "durable-before");
-    assert_eq!(store_inner.dump_journal(&persistence_id).len(), 2);
-}
-
-#[tokio::test]
-async fn dst_reserved_field_update_event_type_cannot_be_dispatched_as_action() {
-    let (_guard, _clock, _id_gen) = install_deterministic_context(18_901);
-    let store_inner = SimEventStore::no_faults(18_901);
-    let store = BoxedEventStore::new(store_inner.clone());
-    let table = Arc::new(RwLock::new(TransitionTable::from_ioa_source(
-        r#"
-[automaton]
-name = "ReservedAction"
-states = ["Draft", "Updated"]
-initial = "Draft"
-
-[[action]]
-name = "$temper.entity.fields-updated.v1"
-kind = "input"
-from = ["Draft"]
-to = "Updated"
-"#,
-    )));
-    let system = ActorSystem::new("dst-reserved-field-update-event");
-    let actor = EntityActor::with_persistence(
-        "ReservedAction",
-        "reserved-action-1",
-        table,
-        serde_json::json!({}),
-        store,
-        BackendLabel::Sim,
-    )
-    .with_tenant("default");
-    let actor_ref = system.spawn(actor, "reserved-action-1");
-
-    let response = dispatch_action(
-        &actor_ref,
-        "$temper.entity.fields-updated.v1",
-        serde_json::json!({}),
-    )
-    .await;
-    assert!(
-        !response.success,
-        "reserved journal type must not run as an action"
-    );
-    assert_eq!(response.state.status, "Draft");
-    let journal = store_inner.dump_journal("default:ReservedAction:reserved-action-1");
-    assert_eq!(journal.len(), 1, "only the bootstrap event may be durable");
-    assert_eq!(journal[0].event_type, "Created");
 }
