+++ crates/temper-server/tests/dst_platform_rollback.rs
@@ -101,6 +101,11 @@ async fn dst_rollback_dispatch_with_store_faults() {
             // Failures are expected — event store faults will cause some to fail.
         }
 
+        // Fault injection covers the dispatch phase. Disable it before recovery
+        // and invariant reads so verification observes the complete durable
+        // journal rather than injecting a new read failure into the checker.
+        let previous_event_faults = faulty_harness.sim_event_store.disable_faults();
+
         // Restart — only successfully persisted state should be visible.
         faulty_harness.restart().await;
 
@@ -120,8 +125,12 @@ async fn dst_rollback_dispatch_with_store_faults() {
             .unwrap_or_else(|e| {
                 panic!(
                     "seed {seed}: data invariants failed after faulty dispatches \
-                     ({success_count} succeeded): {e}"
+                    ({success_count} succeeded): {e}"
                 )
             });
+
+        faulty_harness
+            .sim_event_store
+            .restore_faults(previous_event_faults);
     }
 }
